refactor(renderer): extract FboResources, final draw, worker clock and video graph
